preserve existing clock.dat file content if present

+++ scripts/make_initial_clocks.sh
@@ -7,9 +7,11 @@ pushd $SCRIPT_DIR
 for expt in e21b09 e21e13 e21a14 e21d15 e21a17 e21e18; do
 	for band in band1 band2 band3 band4; do
 		clockfile=../templates/230G/$band/clocks_$expt.dat
-		echo "Making empty $clockfile"
-		echo "CLOCK" > $clockfile
-		echo "# Station ID    Delay (usec)    Rate (usec)" >> $clockfile
+		if [[ ! -f $clockfile ]]; then
+			echo "Making empty $clockfile"
+			echo "CLOCK" > $clockfile
+			echo "# Station ID    Delay (usec)    Rate (usec)" >> $clockfile
+		fi
 	done
 done
 
@@ -17,9 +19,11 @@ done
 for expt in e21f19; do
 	for band in band1 band2 band3 band4; do
 		clockfile=../templates/345G/$band/clocks_$expt.dat
-		echo "Making empty $clockfile"
-		echo "CLOCK" > $clockfile
-		echo "# Station ID    Delay (usec)    Rate (usec)" >> $clockfile
+		if [[ ! -f $clockfile ]]; then
+			echo "Making empty $clockfile"
+			echo "CLOCK" > $clockfile
+			echo "# Station ID    Delay (usec)    Rate (usec)" >> $clockfile
+		fi
 	done
 done
